// src/models/mail-server/mail-server-config.ts
/*************************************************************************
 * Copyright 2021 Gravwell, Inc. All rights reserved.
 * Contact: <legal@gravwell.io>
 *
 * This software may be modified and distributed under the terms of the
 * MIT license. See the LICENSE file for details.
 **************************************************************************/

export interface MailServerConfig {
	insecureSkipVerify: boolean;
	password: string;
	port: number;
	server: string;
	useTLS: boolean;
	username: string;
}