// src/models/mail-server/raw-mail-server-config.ts
/*************************************************************************
 * Copyright 2022 Gravwell, Inc. All rights reserved.
 * Contact: <legal@gravwell.io>
 *
 * This software may be modified and distributed under the terms of the
 * MIT license. See the LICENSE file for details.
 **************************************************************************/

export interface RawMailServerConfig {
	InsecureSkipVerify: boolean;
	Password: string;
	Port: number;
	Server: string;
	UseTLS: boolean;
	Username: string;
}