// src/models/mail-server/raw-mail-server-test-data.ts
/*************************************************************************
 * Copyright 2022 Gravwell, Inc. All rights reserved.
 * Contact: <legal@gravwell.io>
 *
 * This software may be modified and distributed under the terms of the
 * MIT license. See the LICENSE file for details.
 **************************************************************************/

export interface RawMailServerTestData {
	Body: string;
	From: string;
	Subject: string;
	To: Array<string>;
}